Added the controls for the led array
I’ve added in the controls for the led array that displays the RPM
reading. I also moved the OLED_DC pin so it no longer interferes with
the CAN Bus Shield. After this commit, the ArduinoCAN_SD.ino project
will be deprecated. It was a nice thought, but after implementing the
OLED screen, there isn’t enough RAM to power the QueueLists that buffer
the stream before writing to the SD. The main reason, though that it is
being deprecated is that I found out that the ECU already logs data to
an onboard SD card so it would’ve been redundant.